[API][TEST] Refactored out create and login method
