+++ test/api/events/test-event-router.js
@@ -1,0 +1,11 @@
+const mockgoose = require('mockgoose');
+const mongoose = require('mongoose');
+const request = require('supertest');
+const chai = require('chai');
+const sinon = require('sinon');
+
+chai.use(require('sinon-chai'));
+
+const expect = chai.expect;
+
+
